requirements.txt
# ManoMitra - Multilingual Emotional Conversational Agent
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6

# Speech Processing
openai-whisper==20231117
faster-whisper==0.10.0
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.11
sounddevice==0.4.6

# Machine Learning & NLP
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
datasets==2.14.6
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Text-to-Speech
TTS==0.20.6
gTTS==2.4.0

# Audio Analysis & Emotion Recognition
speechbrain==0.5.15
emotion-recognition==0.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
python-jose==3.3.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For advanced features
langchain==0.0.340
sentence-transformers==2.2.2